Made MovieResponseDTO, fixed some previous issues with pom.xml
