+++ .gitignore
@@ -15,7 +15,7 @@ swagger-with-default-responses.json
 .sts4-cache
 
 ### IntelliJ IDEA ###
-.idea
+*.idea
 *.iws
 *.iml
 *.ipr
