+++ .idea/compiler.xml
@@ -7,16 +7,15 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="WeatherTestApi" />
         <module name="WeatherApi" />
       </profile>
     </annotationProcessing>
-    <bytecodeTargetLevel>
-      <module name="WeatherApi" target="1.8" />
-    </bytecodeTargetLevel>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="WeatherApi" options="-parameters" />
+      <module name="WeatherTestApi" options="-parameters" />
     </option>
   </component>
 </project>
