--- WeatherApi/src/main/java/com/hackerrank/weather/model/Genre.java
@@ -1,19 +1,0 @@
-package com.hackerrank.weather.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "GENRE")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Genre {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String name;
-}
