--- WeatherApi/src/main/java/com/hackerrank/weather/repository/GenreRepository.java
@@ -1,7 +1,0 @@
-package com.hackerrank.weather.repository;
-
-import com.hackerrank.weather.model.Genre;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GenreRepository extends JpaRepository<Genre,Integer> {
-}
