--- WeatherApi/src/main/java/com/hackerrank/weather/resource/WeatherRequest.java
@@ -1,33 +1,0 @@
-package com.hackerrank.weather.resource;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-import javax.validation.constraints.NotNull;
-import java.util.Date;
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class WeatherRequest {
-
-    @Temporal(TemporalType.DATE)
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    @NotNull
-    private Date date;
-    @NotNull
-    private Float lat;
-    @NotNull
-    private Float lon;
-    @NotNull
-    private String city;
-    @NotNull
-    private String state;
-    @NotNull
-    private List<Double> temperatures;
-}
