--- WeatherApi/src/main/resources/application.properties
@@ -1,26 +1,0 @@
-server.address=0.0.0.0
-server.port=8000
-
-api.key=0593ba3faa4885b391b88919d9cda544
-
-spring.datasource.url=jdbc:h2:mem:weather_db
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-
-spring.h2.console.enabled=true
-
-spring.jpa.properties.hibernate.show_sql=true
-spring.output.ansi.enabled=always
-spring.mvc.pathmatch.matching-strategy=ant_path_matcher
-spring.jpa.defer-datasource-initialization=true
-
-#server.tomcat.enabled=true
-#server.tomcat.directory=/dev
-#server.tomcat.prefix=stdout
-#server.tomcat.buffered=false
-#server.tomcat.suffix=
-#server.tomcat.file-date-format=
-##server.tomcat.accessLogEnabled=true
-##server.tomcat.accessLogBasedir=/dev
