+++ src/main/java/com/hackerrank/weather/config/SpringFoxConfig.java
