+++ src/main/java/com/hackerrank/weather/controller/WeatherApiRestController.java
@@ -4,6 +4,7 @@ import com.hackerrank.weather.model.Weather;
 import com.hackerrank.weather.resource.WeatherCityAndTempResponse;
 import com.hackerrank.weather.resource.WeatherRequest;
 import com.hackerrank.weather.service.WeatherService;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,7 +16,7 @@ import javax.validation.Valid;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/weather")
@@ -33,8 +34,8 @@ public class WeatherApiRestController {
     }
 
     @GetMapping("/id/{id}")
-    public Weather getWeatherById(@PathVariable("id") Integer id, HttpServletRequest request,HttpServletResponse response) {
-        return weatherService.getWeather(id, request,response);
+    public Weather getWeatherById(@PathVariable("id") Integer id, HttpServletRequest request, HttpServletResponse response) {
+        return weatherService.getWeather(id, request, response);
     }
 
     @GetMapping("/name/{name}")
@@ -54,28 +55,40 @@ public class WeatherApiRestController {
     }
 
     @GetMapping("/city/{cityName}")
-    public ArrayList<Double> getWeatherTemperatureByCityName(@PathVariable(name = "cityName") String cityName){
+    public ArrayList<Double> getWeatherTemperatureByCityName(@PathVariable(name = "cityName") String cityName) {
         return weatherService.getWeatherTemperatureByCityName(cityName);
     }
 
     @DeleteMapping("/delete/id/{id}")
-    public void deleteWeatherById(@PathVariable(name = "id") Integer id){
+    public void deleteWeatherById(@PathVariable(name = "id") Integer id) {
         weatherService.deleteWeatherById(id);
     }
 
     @DeleteMapping("/delete/name/{cityName}")
-    public void deleteFirstInstanceOfWeatherByCity(@PathVariable(name = "cityName")String cityName){
+    public void deleteFirstInstanceOfWeatherByCity(@PathVariable(name = "cityName") String cityName) {
         weatherService.deleteFirstInstanceOfWeatherByCity(cityName);
     }
 
     @GetMapping("/city/{city}/{state}")
-    public List<Weather> getWeathersByCityOrState(@PathVariable(name = "city") String cityName, @PathVariable(name = "state") String stateName){
-        return weatherService.getWeatherS(cityName,stateName);
+    public List<Weather> getWeathersByCityOrState(@PathVariable(name = "city") String cityName, @PathVariable(name = "state") String stateName) {
+        return weatherService.getWeatherS(cityName, stateName);
     }
 
+
     @GetMapping("/temp/{cityName}")
-    public List<WeatherCityAndTempResponse> getTheThirdTempByCityName(@PathVariable(name = "cityName")String cityName){
-       return weatherService.getCityNameAndTemp(cityName);
+    @ApiOperation(value = "Gets all of the temperatures with city name",
+            notes = "Provide a city name to return a list of their temperatures",
+            response = WeatherCityAndTempResponse.class,
+            responseContainer = "List"
+    )
+    public List<WeatherCityAndTempResponse> getTheThirdTempByCityName(@PathVariable(name = "cityName") String cityName) {
+        return weatherService.getCityNameAndTemp(cityName);
+    }
+
+    //removable code
+    @GetMapping("/temp/3/{cityName}")
+    public List<Weather> getTheTempByCity(@PathVariable(name = "cityName") String cityName) {
+        return weatherService.getWeather(cityName);
     }
 
     @ExceptionHandler({MethodArgumentTypeMismatchException.class, NoSuchElementException.class})
