+++ src/main/java/com/hackerrank/weather/model/Genre.java
@@ -1,6 +1,5 @@
 package com.hackerrank.weather.model;
 
-import com.fasterxml.jackson.annotation.JsonAlias;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,14 +7,13 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 
 @Entity
-@Table(name = "genres")
+@Table(name = "GENRE")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Genre {
     @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    @JsonAlias("name")
     private String name;
 }
