+++ src/main/java/com/hackerrank/weather/model/MovieDTOResponse.java
