+++ src/main/java/com/hackerrank/weather/repository/MovieCollectionRepository.java
