+++ src/main/java/com/hackerrank/weather/resource/MovieMapper.java
