+++ src/main/resources/application.properties
@@ -13,3 +13,14 @@ spring.h2.console.enabled=true
 
 spring.jpa.properties.hibernate.show_sql=true
 spring.output.ansi.enabled=always
+spring.mvc.pathmatch.matching-strategy=ant_path_matcher
+spring.jpa.defer-datasource-initialization=true
+
+#server.tomcat.enabled=true
+#server.tomcat.directory=/dev
+#server.tomcat.prefix=stdout
+#server.tomcat.buffered=false
+#server.tomcat.suffix=
+#server.tomcat.file-date-format=
+##server.tomcat.accessLogEnabled=true
+##server.tomcat.accessLogBasedir=/dev
